+++ test/squint/compiler_test.cljs
@@ -478,6 +478,12 @@
   (is (= "bar" (jsv! '(name :foo/bar))))
   (is (= "foo" (jsv! '(namespace :foo/bar))))
   (is (nil? (jsv! '(namespace :foo))))
+  ;; a leading "/" is not a namespace separator: namespace and name use the
+  ;; same threshold so (str (namespace x) "/" (name x)) round-trips
+  (is (nil? (jsv! '(namespace "/one"))))
+  (is (= "/one" (jsv! '(name "/one"))))
+  (is (= "https:" (jsv! '(namespace "https://example.org/all"))))
+  (is (= "/example.org/all" (jsv! '(name "https://example.org/all"))))
   (is (thrown? js/Error (jsv! '(name 42))))
   (is (thrown? js/Error (jsv! '(name nil)))))